crai suggestions
