+++ src/upload.rs
@@ -1,3 +1,5 @@
+use std::env::current_dir;
+use std::path::PathBuf;
 use crate::{
     s3::{get_bucket, get_upload_data},
     upload::machinery::upload_dir_to_bucket,
@@ -6,6 +8,40 @@ use crate::{
 mod machinery;
 
 pub async fn upload(dir: &str) -> color_eyre::Result<()> {
+    let mut failed = false;
+
+    let Ok(dir_path_buffer) = PathBuf::from(&dir).canonicalize() else {
+        eprintln!("unable to canonicalise {}", "[DIR]".blue());
+        std::process::exit(1);
+    };
+    if !dir_path_buffer.exists() {
+        eprintln!("unable to find provided {}", "[DIR]".blue());
+        failed = true;
+    }
+    if !dir_path_buffer.is_dir() {
+        eprintln!("provided {} must be a directory", "[DIR]".blue());
+        failed = true;
+    }
+    match current_dir() {
+        Ok(cd) => {
+            if dir_path_buffer.eq(&cd) {
+                eprintln!(
+                    "provided {} must be a different from current directory",
+                    "[DIR]".blue()
+                );
+                failed = true;
+            }
+        }
+        Err(e) => {
+            eprintln!("unable to access current directory: {e:?}");
+            failed = true;
+        }
+    }
+
+    if failed {
+        std::process::exit(1);
+    }
+
     info!(?dir, "Reading files");
 
     let bucket = get_bucket();
